+++ cxhub_android_firebase/CHANGELOG.md
@@ -1,2 +1,5 @@
 ## 0.0.1-firebase
 * CxHubSDK support.
+
+## 0.0.2-firebase
+* fix native push service onNewToken exception, when sdk is not initialized