+++ cxhub_android_rustore/CHANGELOG.md
@@ -1,2 +1,5 @@
 ## 0.0.1-rustore
 * CxHubSDK support.
+
+## 0.0.2-rustore
+* fix native push service onNewToken exception, when sdk is not initialized